requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
instaloader==4.7.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pafy==0.5.5
Pillow==8.1.2
pycodestyle==2.7.0
python-decouple==3.4
requests==2.25.1
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
youtube-dl==2021.3.14
